Xam.iOS Unified: small fixes so that MG builds
* Texture2D.OpenGL : Making sure we don't cast vars to ints prematurely. CGBitmapContext constructor is expecting nints and although I think this conversion should be occurring implicitly, it appears that one runs into issues if we cast and then pass width/height vars as ints.
* KeyboardInputView: UILabel.StringSize was deprecated and is removed in unified API, so need an alternate approach
* SignedInGamer: Conversion from DateTime to NSDate needs to be explicit. A suggested solution in http://developer.xamarin.com/guides/cross-platform/macios/unified/ is to add extension methods which have been included, but people may prefer for them to be located somewhere else.

+++ MonoGame.Framework/Graphics/Texture2D.OpenGL.cs
@@ -414,8 +414,8 @@ namespace Microsoft.Xna.Framework.Graphics
 #if IOS || MONOMAC
         private static Texture2D PlatformFromStream(GraphicsDevice graphicsDevice, CGImage cgImage)
         {
-			var width = (int)cgImage.Width;
-			var height = (int)cgImage.Height;
+			var width = cgImage.Width;
+			var height = cgImage.Height;
 
             var data = new byte[width * height * 4];
 
@@ -428,7 +428,7 @@ namespace Microsoft.Xna.Framework.Graphics
             Texture2D texture = null;
             Threading.BlockOnUIThread(() =>
             {
-                texture = new Texture2D(graphicsDevice, width, height, false, SurfaceFormat.Color);
+                texture = new Texture2D(graphicsDevice, (int)width, (int)height, false, SurfaceFormat.Color);
                 texture.SetData(data);
             });
